separate storage from abstract.
